refactoring - Task interface removed
