+++ src/test/java/com/example/tasks/TaskReadFromFileTest.java
@@ -6,10 +6,10 @@ import java.util.Collections;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class TaskImplReadFromFileTest {
+public class TaskReadFromFileTest {
     @Test
     public void fileInputTest() {
-        TaskImpl task = new TaskImpl(new String[] {"./src/test/resources/input.txt"}, Collections.emptyList()); // is it file system type save?
+        Task task = new Task(new String[] {"./src/test/resources/input.txt"}, Collections.emptyList()); // is it file system type save?
         task.readInput();
         assertEquals("ONE TWO THREE FOUR ", task.userInput);
     }
